+++ APIproj/GameEngineContents/Monster.h
@@ -22,19 +22,19 @@ public:
 	Monster& operator=(const Monster& _Other) = delete;
 	Monster& operator=(Monster&& _Other) noexcept = delete;
 
-	void Attack(int _Att);
-	inline void SetHp(int _Hp) {
+	void Attack(float _Att);
+	inline void SetHp(float _Hp) {
 		Hp = _Hp;
 	}
 
-	int GetHp() {
+	float GetHp() {
 		return Hp;
 	}
 
-	inline void SetDmg(int _Dmg) {
+	inline void SetDmg(float _Dmg) {
 		Dmg = _Dmg;
 	}
-	inline int GetDmg()
+	inline float GetDmg()
 	{
 		return Dmg;
 	}
@@ -48,8 +48,8 @@ private:
 
 	bool IsAttacked = false;
 	float CollisionOffTime = 0.0f;
-	int Hp = 0;
-	int Dmg = 0;
+	float Hp = 0;
+	float Dmg = 0;
 	float MoveSpeed = 200.0f;
 	float4 MoveVec = float4::Zero;
